Move unused explorts into module that uses them
