--- packages/core/src/js/setInert.js
@@ -1,14 +1,0 @@
-export const setInert = (state, selector) => {
-  if (selector) {
-    const els = document.querySelectorAll(selector);
-    els.forEach((el) => {
-      if (state) {
-        el.inert = true;
-        el.setAttribute('aria-hidden', true);
-      } else {
-        el.inert = null;
-        el.removeAttribute('aria-hidden');
-      }
-    });
-  }
-};
