--- packages/core/src/js/setOverflowHidden.js
@@ -1,12 +1,0 @@
-export const setOverflowHidden = (state, selector) => {
-  if (selector) {
-    const els = document.querySelectorAll(selector);
-    els.forEach((el) => {
-      if (state) {
-        el.style.overflow = 'hidden';
-      } else {
-        el.style.removeProperty('overflow');
-      }
-    });
-  }
-};
